🌱 Extract sub-hooks from useVersionCheck to reduce complexity
Extract pure version-parsing utilities (parseReleaseTag, parseRelease,
getLatestForChannel, isDevVersion, isNewerVersion), JSON helper
(safeJsonParse), localStorage cache helpers, and all related constants
into a new versionUtils.ts file.

The original useVersionCheck.tsx re-exports the public functions so all
existing consumers continue to work without import path changes.

Reduces useVersionCheck.tsx from 1042 to ~800 lines.
